chore: split three files back under their ceilings, and re-measure the suite
CLAUDE.md's ceilings are binding — the two engine crates at 1000 lines, the rest
at 500 — and the work of the last three PRs broke them in three places. All
three are mine, and the worst is the third, because the rule names it exactly:
"new code lands in a small focused module, never appended to something already
oversized".

  parse/item.rs   1345 -> 1395 -> 934   (was ALREADY 345 over, and I added to it)
  ws/client.rs      new -> 504          -> client/{mod 286, url 241}
  ws/conn.rs       436 -> 563           -> conn/{mod 463, reader 115}

The cuts are where the subject changes, not wherever 500 fell:

- `parse/module.rs` is `import` and `export` — one subject, since they are two
  sides of the same table and TypeScript's erasure applies to both.
  `parse/stmt.rs` is what a statement can BE, where `item.rs` keeps what a
  function and a class ARE. The folder already had `expr.rs`, `members.rs` and
  `pat.rs`; these two were the ones missing.
- `ws/client/url.rs` has no socket in it at all — text in, text out — which is
  why its ten tests run with no network, and `client/mod.rs` is everything that
  touches one.
- `ws/conn/reader.rs` is everything that runs on the BACKGROUND thread and may
  never touch JavaScript, where `conn/mod.rs` is the table the JS thread reads.
  The two halves already spoke only through `push`; the file says so now.

What is NOT done: `rts-core` and `rts-codegen` still hold eight files over 1000
(`functions.rs`, `capture.rs`, `objects.rs`, …). Those predate this work and are
untouched — splitting unrelated code to tidy a number is the scope creep the
ceiling rule is not asking for.

AND THE NUMBER IN CLAUDE.md WAS STALE, which rule 7 of `rts-codegen`'s README
calls out by name ("a number quoted from memory is worse than no number"). It
said 758 of 819 for 08-29. Re-measured: **792 of 845** by
`target/release/rts.exe test` at `891b767c`, 2026-09-02.

The paragraph beside it is new and matters more than the count: ONE of the 53
failures is there by design. `generator_for_of_root.test.ts` records an open
defect and used to pass on the build that had it; `expect()` going from eight
allocations per assertion to one stopped hiding it. Anyone reading 792/845
without that reads it wrongly, and the honesty floor is why it is in the file
rather than in a commit message nobody will find.

Verified: `cargo test --profile fast -p rts-node -p rts-codegen -p rts-core`
all green with identical counts, and the suite at 792/845 with an EMPTY lost
list against the measurement before the split.
